Added: Genereting report, mail service, cron service

+++ PowerPlant/PowerPlant.Core/Entities/Production.cs
@@ -5,11 +5,11 @@ namespace PowerPlant.Core.Entities
     public class Production : Entity
     {
         public int GeneratorId { get; set; }
-        public float Quantity { get; set; }
+        public double Quantity { get; set; }
         public DateTime Date { get; set; }
         public TimeSpan Time { get; set; }
 
-        public Production(int generatorId, float quantity)
+        public Production(int generatorId, double quantity)
         {
             GeneratorId = generatorId;
             Quantity = quantity;
@@ -17,7 +17,7 @@ namespace PowerPlant.Core.Entities
             Time = DateTime.Now.TimeOfDay;
         }
 
-        public Production(DateTime date, int generatorId, float quantity)
+        public Production(DateTime date, int generatorId, double quantity)
         {
             GeneratorId = generatorId;
             Quantity = quantity;
